FFLOGS_Dockers/requirements.txt
fastapi==0.94.1
pandas==1.5.2
requests==2.28.2
mysql-connector-python==8.0.32
pymongo==4.3.3
uvicorn==0.21.0